Unify tool PrintUsage and PrintVersion funcs
We had multiple versions of the PrintUsage and PrintVersion
functions, this unifies them.
Additionally, this moves the "global" defines of these from
tool_settings.h which resulted in requiring a #define to filter
them out of some locations, and into a new header called
tool_command_line.h
