+++ android/tools/replay/CMakeLists.txt
@@ -42,6 +42,7 @@ add_library(gfxrecon-replay
 target_include_directories(gfxrecon-replay
                            PUBLIC
                                ${GFXRECON_SOURCE_DIR}/external/precompiled/android/include
+                               ${GFXRECON_SOURCE_DIR}/tools
                                ${GFXRECON_SOURCE_DIR}
                                ${CMAKE_BINARY_DIR})
