+++ tools/file_version_patch/CMakeLists.txt
@@ -32,6 +32,7 @@ target_sources(gfxrecon-file-version-patch
                PRIVATE
                    ${CMAKE_CURRENT_LIST_DIR}/main.cpp
                    ${CMAKE_CURRENT_LIST_DIR}/../platform_debug_helper.cpp
+                   ${CMAKE_CURRENT_LIST_DIR}/../tool_command_line.h
                    $<$<BOOL:WIN32>:${PROJECT_SOURCE_DIR}/version.rc>
                    $<$<BOOL:WIN32>:${GFXRECON_WIN_MANIFEST}>
               )
@@ -46,7 +47,10 @@ if (MSVC)
     endif()
 endif()
 
-target_include_directories(gfxrecon-file-version-patch PUBLIC ${CMAKE_BINARY_DIR})
+target_include_directories(gfxrecon-file-version-patch
+                           PUBLIC
+                              ${CMAKE_CURRENT_LIST_DIR}/..
+                              ${CMAKE_BINARY_DIR})
 target_compile_definitions(gfxrecon_decode PUBLIC VULKAN_HPP_NO_STRUCT_SETTERS)
 
 target_link_libraries(gfxrecon-file-version-patch
