+++ tools/info/CMakeLists.txt
@@ -31,6 +31,7 @@ add_executable(gfxrecon-info "")
 target_sources(gfxrecon-info
                PRIVATE
                    ${CMAKE_CURRENT_LIST_DIR}/../tool_settings.h
+                   ${CMAKE_CURRENT_LIST_DIR}/../tool_command_line.h
                    ${CMAKE_CURRENT_LIST_DIR}/main.cpp
                    ${CMAKE_CURRENT_LIST_DIR}/info_feature.h
                    ${CMAKE_CURRENT_LIST_DIR}/info_d3d12_feature.h
