+++ tools/replay/replay_main_common.h
@@ -28,6 +28,7 @@ struct android_app;
 #endif
 
 #include "replay_feature.h"
+#include "tool_command_line.h"
 
 #include "application/application.h"
 #include "decode/file_processor.h"
