+++ tools/tocpp/CMakeLists.txt
@@ -30,11 +30,15 @@ add_executable(gfxrecon-tocpp "")
 target_sources(gfxrecon-tocpp
                PRIVATE
                    main.cpp
+                   ${CMAKE_CURRENT_LIST_DIR}/../tool_command_line.h
                    $<$<BOOL:WIN32>:${PROJECT_SOURCE_DIR}/version.rc>
                    $<$<BOOL:WIN32>:${GFXRECON_WIN_MANIFEST}>
               )
 
-target_include_directories(gfxrecon-tocpp PUBLIC ${CMAKE_BINARY_DIR})
+target_include_directories(gfxrecon-tocpp
+                           PUBLIC
+                                ${CMAKE_CURRENT_LIST_DIR}/..
+                                ${CMAKE_BINARY_DIR})
 
 target_link_libraries(gfxrecon-tocpp gfxrecon_decode gfxrecon_format gfxrecon_util platform_specific project_version)
